+++ positions/src/modules/index.ts
@@ -1,0 +1,2 @@
+export {PositionToDirective} from "./positions/directives";
+export {PositionsModule} from "./positions/modules";