+++ positions/src/modules/positions/directives/index.ts
@@ -1,0 +1,1 @@
+export * from './positionTo/positionTo.directive';