+++ positions/src/modules/positions/modules/positions.module.ts
@@ -1,0 +1,24 @@
+import {NgModule} from "@angular/core";
+
+import {PositionToDirective} from '../directives';
+
+/**
+ * Module for position to directive
+ */
+@NgModule(
+{
+    imports:
+    [
+    ],
+    declarations:
+    [
+        PositionToDirective
+    ],
+    exports:
+    [
+        PositionToDirective
+    ]
+})
+export class PositionsModule
+{
+}